Use nox and GitHub actions to manage development (#13)
Use GitHub Actions instead of Travis and Azure. Use nox instead of the Makefile
to manage creating environments and running the tests, docs, checks, etc. The
main advantage is that the configuration (noxfile.py) is Python and
cross-platform. Since nox creates the environments, we don't need to install
anything but nox in the conda environment. This simplifies keeping track of
dependencies both locally and on CIs. Had to use conda to build the docs
because cartopy wouldn't install with pip.